+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Collapse duplicate results in Elasticsearch query on uid
 - Add pharmacies to kit pickup for 16-24
 - Replace sexperts search with kit locations for over 25s
+- Use `arc` rather than `plane` for precision in distance
 
 0.5.0 / 2018-04-11
 ==================